[FIX] Proper handling of building interaction
It took me a while to find out why 2 doctors where only providing 1 surgery
It turns out that when a check occurs to look for an interaction in job
(IGenericJob::HasInteraction) it was calling the request (IBuilding::RequestInteraction)
of the building, using the worker instead of just checking for its existence.
The return of RequestInteraction(non null pointer if existing, or null pointer if not)
was perfectly matching what was expected as a boolean, further obfuscating this issue.